+++ src/features/bible/index.ts
@@ -2,3 +2,4 @@ export * from './api';
 export * from './hooks';
 export * from './lib';
 export * from './model';
+export * from './ui';
