+++ src/features/bible/ui/SkeletonCard.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import { Skeleton } from '@/shared/components/ui/skeleton';
+
+function SkeletonCard() {
+  return (
+    <div className="flex flex-col gap-[8px]">
+      {Array.from({ length: 20 }).map((_, i) => (
+        <Skeleton className="h-4 w-full" key={i} />
+      ))}
+    </div>
+  );
+}
+
+export default SkeletonCard;
